Pre-normalized IceSserver objects are passed through.
In the event that an already normalized object is given to the module,
it should return the object as is.

One example case of this is any application using XirSys for STUN/TURN
servers, where you get back an array of already valid IceServer objects.
These, when passed to rtc-io's quickconnect, were getting their
credentials and usernames removed.